# Optimize `HashMap` insertion by removing duplicate computation of hash and position