This PR optimizes `HashMap` (and thus `Dictionary`) insertions by removing duplicate computations.
In particular, `hash` and `pos` only need to be computed once for `insert`.

This change was originally part of https://github.com/godotengine/godot/pull/90082. I have marked @Nazarwadim as a co-author of this PR.
The above PR is very complicated, making it difficult to estimate its effects in their entirety, and mitigate potential for regressions. Therefore, I'm separating them out into small, easy to understand packages.

## Explanation

This is a fairly simple change: Instead of re-computing `hash` for `_lookup_pos` and `_insert`, it is passed as a parameter.
Additionally, it alleviates `_insert` from handling the "key exists" case: One of two callers handled this case already, and the other could be trivially adjusted to do the same. This makes the code cleaner in general, and optimizes both cases.

`_insert_with_hash` is renamed to `_insert_element` to fit its intended purpose better, since `_insert` now also takes a hash parameter.

I do not expect any regressions.

## Benchmarks

I measured an up to 10% performance improvement for `Dictionary` insertion:

```c++
	String s = "Lorem Ipsum dolor sit amed";

	for (int size : { 1, 2, 8, 64, 1024, 4096}) {
		auto t0 = std::chrono::high_resolution_clock::now();

		for (int run = 0; run < 20000000 / size; run++) {
			Dictionary dictionary;
			for (int idx = 0; idx < size; idx ++) {
				// Test
				dictionary.set(idx, s);
			}
		}
		auto t1 = std::chrono::high_resolution_clock::now();

		std::cout << "size:" << size << std::endl;
		std::cout << std::chrono::duration_cast<std::chrono::milliseconds>(t1 - t0).count() << "ms\n";
	}
```

This printed:

```
size:1
2732ms
size:2
2032ms
size:8
1493ms
size:64
1857ms
size:1024
2071ms
size:4096
2167ms
```
(master)

```
size:1
2476ms
size:2
1835ms
size:8
1374ms
size:64
1676ms
size:1024
1868ms
size:4096
2043ms
```
(this PR)

I also measured a 3% improvement in launch time of tps-demo:
```
 ❯ hyperfine -m25 -iw1 "bin/godot.macos.editor.arm64.master --path /Users/lukas/dev/godot/tps-demo --quit" "bin/godot.macos.editor.arm64 --path /Users/lukas/dev/godot/tps-demo --quit"
Benchmark 1: bin/godot.macos.editor.arm64.master --path /Users/lukas/dev/godot/tps-demo --quit
  Time (mean ± σ):      1.523 s ±  0.098 s    [User: 0.577 s, System: 0.184 s]
  Range (min … max):    1.436 s …  1.868 s    25 runs

Benchmark 2: bin/godot.macos.editor.arm64 --path /Users/lukas/dev/godot/tps-demo --quit
  Time (mean ± σ):      1.484 s ±  0.039 s    [User: 0.572 s, System: 0.181 s]
  Range (min … max):    1.433 s …  1.598 s    25 runs

Summary
  bin/godot.macos.editor.arm64 --path /Users/lukas/dev/godot/tps-demo --quit ran
    1.03 ± 0.07 times faster than bin/godot.macos.editor.arm64.master --path /Users/lukas/dev/godot/tps-demo --quit
```