@clayjohn 2025-05-09
Do you have a theory for why the single element case got more expensive with this change? Perhaps there is more overhead in creating the dictionary now?