@lawnjelly 2025-05-10
> Do you have a theory for why the single element case got more expensive with this change? Perhaps there is more overhead in creating the dictionary now?

I think they all look faster, maybe I didn't spot this case (the table in the OP was kind of confusing because the title was _after_ the figures).