@Ivorforce 2025-05-10
Yes, all cases are faster with this change :)